Add rhai scripting

+++ crates/core/src/settings.rs
@@ -5,13 +5,13 @@ use std::borrow::Cow;
 )]
 #[allow(non_camel_case_types)]
 pub struct FactorioSettings {
-    pub client_count: i64,
+    pub client_count: u32,
     pub factorio_archive_path: Cow<'static, str>,
     pub map_exchange_string: Cow<'static, str>,
     pub rcon_pass: Cow<'static, str>,
-    pub rcon_port: i64,
+    pub rcon_port: u16,
     pub recreate: bool,
-    pub restapi_port: i64,
+    pub restapi_port: u16,
     pub seed: Cow<'static, str>,
     pub workspace_path: Cow<'static, str>,
 }
